+++ agentshield/checks/__init__.py
@@ -1,3 +1,4 @@
+from .agent_tools import scan_agent_tools
 from .environment import scan_environment
 from .git_config import scan_git_config
 from .global_packages import scan_global_packages
@@ -7,10 +8,10 @@ from .ssh import scan_ssh
 
 __all__ = [
     "scan_environment",
+    "scan_agent_tools",
     "scan_git_config",
     "scan_global_packages",
     "scan_secret_files",
     "scan_shell_history",
     "scan_ssh",
 ]
-